Remove "color/c/hex" abbreviation
Now formatted like "<#123456>"
