+++ docs/text.html
@@ -112,19 +112,8 @@ Here's a table of all colors, including both Skript names and color codes:
 </table>
 
 In Minecraft 1.16, support was added for 6-digit hexadecimal colors to specify custom colors other than the 16 default color codes.
-A new tag was added to allow usage of these colors:
-<table>
-  <tr>
-    <td>Name</td>
-    <td>Alternative Names</td>
-    <td>Description</td>
-  </tr>
-  <tr>
-    <td>color</td>
-    <td>c, hex</td>
-    <td>Sets the color of the message based on a 6-digit hexadecimal color</td>
-  </tr>
-</table>
+A new tag can be used to format with these colors. The tag looks like "<#hex code>".<br>
+Here's what the tag would look like when used in a script: send "<#123456>Hey %player%!" to player<br><br>
 
 For information not related to Skript, see
 <a href="https://minecraft.gamepedia.com/Formatting_codes#Color_codes">Minecraft
