+++ src/main/resources/lang/english.lang
@@ -524,7 +524,6 @@ colors:
 
 # -- Chat Styles --
 chat styles:
-	color: color, c, hex
 	bold: bold, b
 	italic: italic, italics, i
 	strikethrough: strikethrough, strike, s
